+++ src/Symfony/Component/Mailer/Bridge/Brevo/CHANGELOG.md
@@ -4,6 +4,8 @@ CHANGELOG
 8.2
 ---
 
+* Add `TemplateParametersHeader` to pass Brevo template parameters with arbitrary types
+  (string, int, bool, array) instead of the generic `ParameterizedHeader` which only accepts strings
 * Allow configuring the SMTP port through the DSN
 
 6.4